# Optimizations on System.Linq.Enumerable.ToArray