… src/libraries/System.Linq/src/System/Linq/Select.SpeedOpt.cs