3rd of n optimizations lifted from my exploration of an alternative `System.Linq` implementation [`Cistern.Linq`](https://github.com/manofstick/Cistern.Linq). There are lots more goodies there, but only so many that will survive the transfer to `System.Linq`!

([_Previous corefx optimization_](https://github.com/dotnet/runtime/pull/336))

This is a bit more controversial, which although I think it is reasonable, may get kicked to the curb depending on you POV. So why do I think this is is controversial? Well it uses a [Clayton's](https://en.wikipedia.org/wiki/Claytons) `stackalloc` (i.e. recursion) to store the first 64 elements (sounded like a reasonable number, but obviously up for discussion) to avoid temporary allocation buffers (can't use a "real" `stackalloc` due to not meeting the generic type constraints).

It then finishes off the process using recursion too. What this means is that you have got some deeper stack accesses than previously, but _I feel_ that they are not unreasonable. So how much stack are we using?

So the maximum function depth is 16 + 25 = 41

max array size ~ 2^31

approx stack usage = 16*(4*sizeof(T)+sizeof(T[])+sizeof(IEnumerable<>)+sizeof(int)+sizeof(Func<>)+sizeof(int)+sizeof(this)+function overhead) +  (log2(2^31/64))*(sizeof(object)+sizeof(ref int)+sizeof(Func<>)+sizeof(count)+function overhead).

which, for a sizeof(T) = 8 is ~ 2K

Anyway, so if that is unreasonable, then you can stop here and kill this PR. Otherwise, read on...

So I added some ToArray performance tests, but haven't pushed them to the Performance repository (as they are a bit excessive for what they are doing there), but you can [view them here](https://github.com/manofstick/performance/blob/9e568cc245d13416a044995013640ed4d393db9a/src/benchmarks/micro/corefx/System.Linq/Perf.ToArray.cs).

The tests are modifying 3 axes... Length, the type of predicate used in `Where` and the source type (i.e. a generator function, an array or a list). What you are not seeing here is the smaller amount of allocation that are occuring due to avoiding temporary buffers for small arrays.

Anyway, I'll add some more details - some memory usage, etc. (and I realised that I haven't actually included the case of the Generator without a `Where`) if this is actually going to be considered to be pulled into `master`

(oh, and the plan would also be to extend this to WhereSelect as well, but TBD...)

 Slower                                                                        | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| ----------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| -------- |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 0, Pred: FirstHalf, Enum: Array)     |      1.15 |            15.34 |            17.64 | several?|
| System.Linq.Tests.Perf_ToArray.Linq(Len: 10000, Pred: FirstHalf, Enum: Array) |      1.08 |         32662.74 |         35416.31 |         |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 0, Pred: Interleaved, Enum: Array)   |      1.05 |            16.57 |            17.32 |         |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 500, Pred: False, Enum: Array)       |      1.04 |          1376.24 |          1429.36 |         |

| Faster                                                                           | base/diff | Base Median (ns) | Diff Median (ns) | Modality  |
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| ---------- |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 12, Pred: True, Enum: Array)            |      2.33 |           182.54 |            78.49 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 12, Pred: True, Enum: List)             |      2.29 |           182.99 |            79.78 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 25, Pred: True, Enum: List)             |      2.29 |           282.44 |           123.43 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 2, Pred: FirstHalf, Enum: List)         |      2.27 |           101.11 |            44.63 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 12, Pred: FirstHalf, Enum: Array)       |      2.18 |           158.58 |            72.82 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 3, Pred: Interleaved, Enum: List)       |      2.18 |           103.74 |            47.68 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 3, Pred: FirstHalf, Enum: List)         |      2.16 |           104.79 |            48.56 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 25, Pred: True, Enum: Array)            |      2.15 |           265.49 |           123.35 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 2, Pred: FirstHalf, Enum: Array)        |      2.14 |            98.21 |            45.87 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 2, Pred: Interleaved, Enum: List)       |      2.12 |           106.41 |            50.29 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 12, Pred: FirstHalf, Enum: List)        |      2.08 |           154.72 |            74.26 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 6, Pred: True, Enum: Array)             |      2.06 |           114.62 |            55.68 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 6, Pred: Interleaved, Enum: Array)      |      2.04 |           112.34 |            55.01 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 3, Pred: FirstHalf, Enum: Array)        |      2.02 |           103.16 |            51.06 | several?  |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 12, Pred: Interleaved, Enum: List)      |      2.02 |           161.18 |            79.86 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 6, Pred: FirstHalf, Enum: List)         |      2.02 |           115.88 |            57.49 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 2, Pred: Interleaved, Enum: Array)      |      2.01 |           101.01 |            50.32 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 6, Pred: True, Enum: List)              |      1.99 |           118.19 |            59.34 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 6, Pred: Interleaved, Enum: List)       |      1.96 |           114.94 |            58.51 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 3, Pred: Interleaved, Enum: Array)      |      1.96 |           104.32 |            53.35 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 6, Pred: FirstHalf, Enum: Array)        |      1.95 |           112.09 |            57.41 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 2, Pred: True, Enum: Array)             |      1.94 |            79.82 |            41.24 | several?  |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 1, Pred: Interleaved, Enum: Generator)  |      1.91 |           130.29 |            68.07 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 51, Pred: True, Enum: List)             |      1.90 |           425.07 |           224.20 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 1, Pred: True, Enum: Generator)         |      1.84 |           121.67 |            66.02 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 1, Pred: True, Enum: Array)             |      1.83 |            76.45 |            41.66 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 25, Pred: Interleaved, Enum: List)      |      1.82 |           238.43 |           131.28 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 12, Pred: Interleaved, Enum: Array)     |      1.81 |           149.81 |            82.90 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 51, Pred: FirstHalf, Enum: List)        |      1.80 |           370.30 |           205.91 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 3, Pred: Interleaved, Enum: Generator)  |      1.79 |           145.91 |            81.32 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 1, Pred: True, Enum: List)              |      1.79 |            76.82 |            42.82 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 25, Pred: FirstHalf, Enum: Array)       |      1.79 |           225.20 |           125.55 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 1, Pred: Interleaved, Enum: List)       |      1.79 |            77.34 |            43.17 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 2, Pred: True, Enum: List)              |      1.79 |            81.51 |            45.59 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 3, Pred: True, Enum: Array)             |      1.78 |            83.51 |            46.84 | several?  |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 51, Pred: FirstHalf, Enum: Array)       |      1.78 |           344.55 |           193.87 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 6, Pred: True, Enum: Generator)         |      1.77 |           188.60 |           106.76 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 3, Pred: True, Enum: List)              |      1.76 |            85.55 |            48.71 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 25, Pred: Interleaved, Enum: Array)     |      1.75 |           217.37 |           124.39 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 2, Pred: True, Enum: Generator)         |      1.75 |           133.35 |            76.33 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 2, Pred: Interleaved, Enum: Generator)  |      1.75 |           128.81 |            73.80 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 25, Pred: FirstHalf, Enum: List)        |      1.74 |           221.55 |           127.22 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 51, Pred: Interleaved, Enum: List)      |      1.72 |           393.70 |           228.39 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 1, Pred: Interleaved, Enum: Array)      |      1.72 |            75.78 |            44.16 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 3, Pred: FirstHalf, Enum: Generator)    |      1.70 |           143.33 |            84.53 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 51, Pred: True, Enum: Array)            |      1.69 |           367.21 |           217.87 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 25, Pred: True, Enum: Generator)        |      1.68 |           462.87 |           275.08 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 3, Pred: True, Enum: Generator)         |      1.68 |           139.65 |            83.00 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 51, Pred: Interleaved, Enum: Array)     |      1.66 |           355.92 |           214.32 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 100, Pred: Interleaved, Enum: List)     |      1.64 |           643.61 |           393.09 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 2, Pred: FirstHalf, Enum: Generator)    |      1.62 |           130.37 |            80.43 | bimodal   |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 100, Pred: FirstHalf, Enum: Array)      |      1.59 |           549.81 |           346.58 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 6, Pred: FirstHalf, Enum: Generator)    |      1.58 |           164.40 |           104.17 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 100, Pred: True, Enum: List)            |      1.57 |           694.40 |           442.13 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 12, Pred: Interleaved, Enum: Generator) |      1.56 |           246.04 |           157.64 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 1, Pred: False, Enum: Array)            |      1.55 |            57.20 |            36.89 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 2, Pred: False, Enum: Array)            |      1.54 |            59.82 |            38.74 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 51, Pred: Interleaved, Enum: Generator) |      1.54 |           734.28 |           477.95 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 6, Pred: Interleaved, Enum: Generator)  |      1.53 |           166.31 |           108.52 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 25, Pred: Interleaved, Enum: Generator) |      1.53 |           405.54 |           264.73 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 100, Pred: FirstHalf, Enum: List)       |      1.53 |           576.73 |           376.60 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 100, Pred: Interleaved, Enum: Array)    |      1.53 |           567.35 |           371.47 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 51, Pred: True, Enum: Generator)        |      1.52 |           757.28 |           496.94 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 12, Pred: True, Enum: Generator)        |      1.52 |           265.65 |           175.29 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 0, Pred: True, Enum: List)              |      1.51 |            51.22 |            33.88 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 0, Pred: False, Enum: List)             |      1.51 |            51.06 |            33.87 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 0, Pred: Interleaved, Enum: List)       |      1.50 |            51.32 |            34.11 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 100, Pred: True, Enum: Array)           |      1.49 |           622.51 |           417.38 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 3, Pred: False, Enum: List)             |      1.49 |            65.20 |            43.74 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 12, Pred: FirstHalf, Enum: Generator)   |      1.49 |           235.21 |           157.86 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 0, Pred: False, Enum: Generator)        |      1.49 |            80.73 |            54.25 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 0, Pred: FirstHalf, Enum: Generator)    |      1.48 |            77.13 |            52.17 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 1, Pred: FirstHalf, Enum: Array)        |      1.48 |            58.28 |            39.46 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 0, Pred: FirstHalf, Enum: List)         |      1.47 |            51.81 |            35.15 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 2, Pred: False, Enum: List)             |      1.45 |            58.91 |            40.73 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 51, Pred: FirstHalf, Enum: Generator)   |      1.40 |           668.98 |           477.35 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 100, Pred: Interleaved, Enum: Generator |      1.39 |          1249.75 |           897.68 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 0, Pred: True, Enum: Generator)         |      1.38 |            73.75 |            53.36 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 1, Pred: FirstHalf, Enum: Generator)    |      1.38 |            80.73 |            58.46 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 1, Pred: False, Enum: List)             |      1.38 |            52.82 |            38.32 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 1, Pred: FirstHalf, Enum: List)         |      1.36 |            54.37 |            40.03 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 500, Pred: True, Enum: List)            |      1.36 |          2421.67 |          1783.85 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 10000, Pred: Interleaved, Enum: List)   |      1.36 |         45226.34 |         33325.68 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 6, Pred: False, Enum: Array)            |      1.35 |            66.03 |            48.81 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 25, Pred: FirstHalf, Enum: Generator)   |      1.35 |           372.16 |           275.16 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 6, Pred: False, Enum: List)             |      1.35 |            71.74 |            53.25 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 500, Pred: Interleaved, Enum: List)     |      1.34 |          2370.59 |          1765.25 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 1, Pred: False, Enum: Generator)        |      1.33 |            80.45 |            60.29 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 100, Pred: True, Enum: Generator)       |      1.33 |          1271.30 |           955.42 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 1000, Pred: True, Enum: List)           |      1.32 |          4562.01 |          3455.84 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 12, Pred: False, Enum: List)            |      1.30 |            92.53 |            71.24 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 0, Pred: Interleaved, Enum: Generator)  |      1.30 |            77.81 |            60.04 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 2, Pred: False, Enum: Generator)        |      1.27 |            87.50 |            68.98 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 3, Pred: False, Enum: Generator)        |      1.25 |            95.61 |            76.75 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 25, Pred: False, Enum: List)            |      1.24 |           148.23 |           119.17 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 3, Pred: False, Enum: Array)            |      1.24 |            58.82 |            47.30 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 1000, Pred: Interleaved, Enum: List)    |      1.24 |          4533.83 |          3666.57 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 500, Pred: True, Enum: Generator)       |      1.23 |          5217.96 |          4236.11 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 500, Pred: FirstHalf, Enum: List)       |      1.23 |          1965.41 |          1597.56 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 100, Pred: FirstHalf, Enum: Generator)  |      1.23 |          1090.96 |           890.22 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 1000, Pred: FirstHalf, Enum: List)      |      1.23 |          3761.06 |          3069.51 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 10000, Pred: True, Enum: List)          |      1.22 |         41794.84 |         34157.86 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 100, Pred: False, Enum: List)           |      1.21 |           416.73 |           344.04 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 51, Pred: False, Enum: List)            |      1.21 |           241.05 |           199.08 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 1000, Pred: True, Enum: Generator)      |      1.20 |         10135.89 |          8451.64 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 6, Pred: False, Enum: Generator)        |      1.20 |           120.48 |           100.47 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 500, Pred: True, Enum: Array)           |      1.20 |          1956.78 |          1636.89 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 10000, Pred: FirstHalf, Enum: List)     |      1.18 |         35312.14 |         29887.81 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 10000, Pred: True, Enum: Generator)     |      1.18 |         97575.66 |         83007.15 |           |
| System.Linq.Tests.Perf_ToArray.Linq(Len: 12, Pred: False, Enum: Generator)       |      1.18 |           179.29 |           152.55 |           |
…