@manofstick 2019-12-05
...and some memory usage...

| Method | Len | Pred |  Enum | Allocated (original) | Allocated (this PR) | ratio |
|------- |---- |----- |------ |----------:|----------:|------:|
|   Linq |   0 | True | Array |         - |         - |       |
|   Linq |   1 | True | Array |      80 B |      80 B |  1.0  |
|   Linq |   3 | True | Array |      88 B |      88 B |  1.0  |
|   Linq |   9 | True | Array |     112 B |     264 B |  0.42 |
|   Linq |  27 | True | Array |     184 B |     464 B |  0.40 |
|   Linq |  81 | True | Array |     680 B |     944 B |  0.72 |
|   Linq | 243 | True | Array |    1864 B |    2264 B |  0.82 |