@stephentoub 2019-12-07
Thanks, @manofstick.

I've not looked at the changes yet, but I am concerned by:

> So why do I think this is is controversial? Well it uses a Clayton's stackalloc (i.e. recursion) to store the first 64 elements

This is effectively saying the code is doing `stackalloc T[64]` (even though you can't actually write that, which is why you're employing recursion).  I don't think we'd ever choose to do that in a library like this, because we have no idea what T is or how large it is.  It could be a very large struct, in which case this is a recipe for a stack overflow that kills the process, all because Enumerable.ToArray was used.