@manofstick 2019-12-07
@stephentoub

> This is effectively saying the code is doing stackalloc T[64] (even though you can't actually write that, which is why you're employing recursion).

Yes, but worse (due to other overheads)! :-)

> I don't think we'd ever choose to do that in a library like this, because we have no idea what T is or how large it is. It could be a very large struct, in which case this is a recipe for a stack overflow that kills the process, all because Enumerable.ToArray was used.

Given that we have a 1Mb stack, if were from the root then that would be structs of ~16K. Pretty big. OK, can't expect that we'd be from the root, so 1/2 Mb - already a *hugely* deep call stack, then we ~8 K? I think that's a stretch too... But obviosuly you have to be conservative in a library. I understand, I'm not trying to argue *too* much :-), and obviously you're call is the law, but is there any size that we can consider as reasonable?