@manofstick 2019-12-10
@stephentoub 

OK, how's this for an idea...

First of all I can do (array|list).Where.ToArray() without storing T, as I can just store indexes. So these should work all fine.

Second I can do quick value type/ref type check (i.e `default is null`) so I can the recurse with ref types, which should be all fine.

Third for other types I just go through my recursive store only once - i.e. storing only the first 4 values. Now this is safe (ish) as the largest value type that can be used in an array is 128k (the Type Loader throws an exception for bigger sizes from my testing....), so at most I consume - in the _exceedingly unlikely_ case - only half the stack. But really I can't see anyone realistically created Value Types anywhere near this, and if they did, they are likely blowing the stack elsewhere, or have increased it's size anyway... 

Anyway, if you still think this is all just crazy talk, do feel free to shoot me down. I've got a pretty thick skin :-)