@stephentoub 2020-03-02
@manofstick, apologies for the long delay in getting back to this.

I spent some time looking at the change.  If I can summarize what you're trying to do here, it's to cater to small ToArray inputs by unrolling the loop and using stack allocation for the temporary values before they're stored into the small result array, yes?

It's an interesting approach.  I just don't think it's going to move the needle sufficiently to warrant its weight (it also drops some existing optimizations for things like e.g. Enumerable.Range(...).ToArray), especially once the recursion / stackoverflow concerns are addressed and the limit drops to four entries as you suggested.  Using an approach like @jkotas did in https://github.com/dotnet/runtime/pull/32365#issuecomment-586648783, I took a look at the overhead this incurs just in generic instantations, and for a simple case, it was increasing working set by ~2.5K per generic instantiation; even for the "speed optimized" build, that's hefty for such a commonly-used method as ToArray.

FWIW, I actually think we should be looking to go in the other direction and simplify the implementation of ToArray we already have.  We accepted the contributions to get it to where it is, but all of the builder stuff that's trying to improve memory usage and throughput is adding a lot of complexity and also itself significantly increasing working set.  I just tried replacing the three lines at https://github.com/dotnet/runtime/blob/0bf721ab4dcd3426790c3251788f77262e8abb1c/src/libraries/Common/src/System/Collections/Generic/EnumerableHelpers.Linq.cs#L110-L112 with a super simple "enumerate the source, store the items into a growing array, and return a resized array result" implementation, and it reduced working set by more than this PR is increasing it, appx 4.5K per generic instantiation!

I'll defer to @eiriktsarpalis and @adamsitnik, but I think we should close this PR, but keep some of your ideas in mind as we look at ways to simplify what's currently there; maybe we can have our cake and eat it, too.  At the moment, though, I don't think this yields the right trade-offs.

Thanks!