@manofstick 2020-03-10
@stephentoub

Thanks for taking the time to peruse. This one always was a bit of a long-shot, but I'm glad I threw it out there and hopefully it will seed some ideas for you.

I'll close it off.